Simplified education endpoints.
