--- src/app/api/resume/[candidateId]/education/credentials/route.ts
@@ -1,73 +1,0 @@
-// Dependencies.
-import { type NextRequest, NextResponse } from "next/server"
-import { authorizeApiToken } from "@/lib/api/auth"
-import { credentialCreateSchema } from "@/lib/api/schemas/resume/education/contract"
-import {
-	parseJson,
-	validateDataFound,
-	validateRequestBodyAgainstSchema,
-	validateUuidFormat,
-} from "@/lib/api/validate"
-import { findCandidateByCandidateId } from "@/lib/db/resume/candidate/sql"
-import { createCredentialByCandidateId } from "@/lib/db/resume/education/sql"
-
-//
-// POST /api/resume/[candidateId]/education/credentials.
-//
-export async function POST(
-	request: NextRequest,
-	{ params }: { params: Promise<{ candidateId: string }> },
-) {
-	// Candidate ID.
-	const { candidateId } = await params
-
-	// If the candidate ID isn’t a valid UUID, return 400.
-	const uuidFormatErrorResponse = validateUuidFormat([candidateId])
-	if (uuidFormatErrorResponse) return uuidFormatErrorResponse
-
-	// If authorization fails, return 401, 403, or 404.
-	const authorizationErrorResponse = await authorizeApiToken(request, {
-		candidateId,
-		scopeRequirement: "resume:write",
-	})
-	if (authorizationErrorResponse) return authorizationErrorResponse
-
-	// Found candidate.
-	const foundCandidate = await findCandidateByCandidateId(candidateId)
-
-	// If the candidate’s not found, return 404.
-	const candidateErrorResponse = validateDataFound(
-		foundCandidate,
-		"candidate",
-		{ candidateId },
-	)
-	if (candidateErrorResponse) return candidateErrorResponse
-
-	// If parsing the request body fails, return 400.
-	const requestBodyOrErrorResponse = await parseJson(request)
-	if (requestBodyOrErrorResponse instanceof NextResponse)
-		return requestBodyOrErrorResponse
-
-	// Request body.
-	const requestBody = requestBodyOrErrorResponse
-
-	// If validating the request body against the schema fails, return 400.
-	const validatedRequestBodyOrErrorResponse = validateRequestBodyAgainstSchema(
-		requestBody,
-		credentialCreateSchema,
-	)
-	if (validatedRequestBodyOrErrorResponse instanceof NextResponse)
-		return validatedRequestBodyOrErrorResponse
-
-	// Validated request body.
-	const validatedRequestBody = validatedRequestBodyOrErrorResponse
-
-	// Created credential.
-	const createdCredential = await createCredentialByCandidateId(
-		candidateId,
-		validatedRequestBody,
-	)
-
-	// If the credential’s created, return 201.
-	return NextResponse.json({ credential: createdCredential }, { status: 201 })
-}
